Update: Tried to fix learning issue
Architecture was updated and changed according to the proposal
in the paper. However, the accuracy after learning is
as high as guessing by chance. It could not be figured out
why this is the case. Unfortunately, since keras is a high
level api it is hard to figure out what happens when.