Update: Removed while loop and cleaned code
Since the cuda implementation has difficulties analysing
the tesnforflow while loop properly, it was replaced by
the previous for loop implementation. After checking
the tensorflow graph it is assumed that the gradient computation
and the forward pass work similarly and give the same results.

Moreover, the code has been cleaned up-